+++ src/Controller/FilmController.php
@@ -65,7 +65,6 @@ class FilmController extends AbstractController
     {
         $newFilm = $serializer->deserialize($request->getContent(), Film::class, 'json');
         $currentFilm->setTitre($newFilm->getTitre());
-        $currentFilm->setDate($newFilm->getDate());
         $currentFilm->setSortieFilm($newFilm->getSortieFilm());
         $currentFilm->setImdb($newFilm->getImdb());
