+++ src/Entity/Proposition.php
@@ -30,6 +30,10 @@ class Proposition
     #[Groups(["filmsGagnants"])]
     private ?Semaine $semaine = null;
 
+    #[ORM\Column(type: 'datetime_immutable', nullable: false)]
+    #[Groups(["getPropositions", "filmsGagnants"])]
+    private ?\DateTimeInterface $date = null;
+
     public function getId(): ?int
     {
         return $this->id;
@@ -71,4 +75,15 @@ class Proposition
         return $this;
     }
 
+    public function getDate(): ?\DateTimeInterface
+    {
+        return $this->date;
+    }
+
+    public function setDate(\DateTimeInterface $date): self
+    {
+        $this->date = $date;
+
+        return $this;
+    }
 }
